Some small bug

+++ RobotClient/Models/FocusModel.cs
@@ -5,7 +5,7 @@ namespace RobotInterface.Models
     public class FocusModel
     {
         public string Name { get; set; }
-        public int Val { get; set; }
+        public double Val { get; set; }
         public int Idx { get; set; }
     }
 }
